backend/Exam_management/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
django-cors-headers==4.3.0
django-filter==23.3
Pillow==10.1.0
qrcode==7.4.2
python-decouple==3.8
